pyproject.toml
[tool.poetry]
name = "clip-classifier"
version = "0.1.0"
description = ""
authors = ["mvalente <miguelvalente@protonmail.com>"]
readme = "README.md"
packages = [{include = "clip_classifier"}]

[tool.poetry.dependencies]
python = "^3.10"
ftfy = "^6.1.1"
regex = "^2022.10.31"
tqdm = "^4.64.1"
clip = {git = "https://github.com/openai/CLIP.git"}
fastapi = "^0.90.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
python-multipart = "^0.0.5"
torch = "^1.13.1"
pillow = "^9.4.0"
requests = "^2.28.2"
torchvision = "^0.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
